Move karma config into karma-config package
