+++ packages/karma-config/README.md
@@ -1,0 +1,5 @@
+# Karma Config
+
+## Overview
+
+Shared Karma configuration for projects in the Nimble repo
