+++ packages/karma-config/index.js
@@ -1,0 +1,80 @@
+// Based on fast-components configuration:
+// https://github.com/microsoft/fast/blob/6549309c1ed2dea838561d23fea6337ef16d7908/packages/web-components/fast-components/karma.conf.js
+// Coverage from the fast configuration removed due to lack of Webpack 5 support:
+// https://github.com/webpack-contrib/istanbul-instrumenter-loader/issues/110
+
+const playwright = require('playwright');
+
+process.env.WEBKIT_HEADLESS_BIN = playwright.webkit.executablePath();
+process.env.WEBKIT_BIN = playwright.webkit.executablePath();
+process.env.FIREFOX_BIN = playwright.firefox.executablePath();
+process.env.CHROME_BIN = playwright.chromium.executablePath();
+
+const commonChromeFlags = [
+    '--no-default-browser-check',
+    '--no-first-run',
+    '--no-sandbox',
+    '--no-managed-user-acknowledgment-check',
+    '--disable-background-timer-throttling',
+    '--disable-backing-store-limit',
+    '--disable-boot-animation',
+    '--disable-cloud-import',
+    '--disable-contextual-search',
+    '--disable-default-apps',
+    '--disable-extensions',
+    '--disable-infobars',
+    '--disable-translate',
+    '--force-prefers-reduced-motion',
+    '--lang=en-US',
+    '--time-zone-for-testing=America/Chicago'
+];
+
+module.exports = config => {
+    const options = {
+        browserDisconnectTimeout: 10000,
+        processKillTimeout: 10000,
+        frameworks: [
+            'jasmine'
+        ],
+        plugins: [
+            'karma-jasmine',
+            'karma-jasmine-html-reporter',
+            'karma-chrome-launcher',
+            'karma-firefox-launcher',
+            'karma-webkit-launcher'
+        ],
+        mime: {
+            'text/x-typescript': ['ts']
+        },
+        reporters: ['kjhtml'],
+        browsers: ['ChromeHeadlessOpt'],
+        customLaunchers: {
+            ChromeDebugging: {
+                base: 'Chrome',
+                flags: [...commonChromeFlags, '--remote-debugging-port=9333'],
+                debug: true
+            },
+            ChromeHeadlessOpt: {
+                base: 'ChromeHeadless',
+                flags: [...commonChromeFlags]
+            },
+            FirefoxDebugging: {
+                base: 'Firefox',
+                debug: true
+            },
+            WebkitDebugging: {
+                base: 'Webkit',
+                debug: true
+            }
+        },
+        client: {
+            jasmine: {
+                stopSpecOnExpectationFailure: false
+            },
+            captureConsole: true
+        },
+        logLevel: config.LOG_ERROR // to disable the WARN 404 for image requests
+    };
+
+    config.set(options);
+};
