+++ packages/karma-config/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@ni-private/karma-config",
+  "version": "1.0.0",
+  "private": true,
+  "type": "commonjs",
+  "description": "Shared Karma configuration for projects in the Nimble repo",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ni/nimble.git",
+    "directory": "packages/karma-config"
+  },
+  "author": "National Instruments",
+  "license": "MIT",
+  "peerDependencies": {
+    "jasmine": "^5.1.0",
+    "jasmine-core": "^5.1.2",
+    "karma": "^6.3.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-firefox-launcher": "^2.1.0",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "karma-jasmine-spec-tags": "^2.0.0",
+    "karma-source-map-support": "^1.4.0",
+    "karma-sourcemap-loader": "^0.4.0",
+    "karma-webkit-launcher": "^2.1.0",
+    "karma-webpack": "^5.0.0",
+    "playwright": "1.42.0",
+    "webpack": "^5.75.0"
+  },
+  "files": ["index.js", "webpack.js"],
+  "exports": {
+    ".": "./index.js",
+    "./webpack": "./webpack.js",
+    "./webpack.js": "./webpack.js"
+  }
+}
