+++ packages/nimble-components/package.json
@@ -113,6 +113,7 @@
     "@ni/eslint-config-javascript": "^4.2.0",
     "@ni/eslint-config-typescript": "^4.2.0",
     "@ni/jasmine-parameterized": "^0.2.4",
+    "@ni-private/karma-config": "^1.0.0",
     "@rollup/plugin-commonjs": "^25.0.7",
     "@rollup/plugin-json": "^6.0.0",
     "@rollup/plugin-node-resolve": "^15.0.1",
@@ -140,20 +141,8 @@
     "eslint-plugin-jsdoc": "^48.2.0",
     "eslint-plugin-storybook": "^0.8.0",
     "html-webpack-plugin": "^5.3.1",
-    "jasmine-core": "^5.1.2",
     "js-beautify": "^1.15.1",
-    "karma": "^6.3.0",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-firefox-launcher": "^2.1.0",
-    "karma-jasmine": "^5.1.0",
-    "karma-jasmine-html-reporter": "^2.0.0",
-    "karma-jasmine-spec-tags": "^2.0.0",
-    "karma-source-map-support": "^1.4.0",
-    "karma-sourcemap-loader": "^0.4.0",
     "karma-spec-reporter": "^0.0.36",
-    "karma-webkit-launcher": "^2.1.0",
-    "karma-webpack": "^5.0.0",
-    "playwright": "1.42.0",
     "prettier-eslint": "^16.3.0",
     "prettier-eslint-cli": "^8.0.1",
     "remark-gfm": "^4.0.0",
